Move dispatcher-event headers out of app/components/ (refs #1194)
audio_events.h, haptics.h, and input_events.h are dispatcher payloads
that live on entt::dispatcher, never on entities. Per the app/components/
audit (#1194), they don't pass the atomic-and-queryable component test
and belong beside their owning systems.

Move all three headers from app/components/ to app/systems/. The earlier
app/audio/ and app/input/ folders cited in the audit no longer exist
(collapsed in #1216/#1217), so app/systems/ is the canonical home.

- app/components/audio_events.h -> app/systems/audio_events.h
- app/components/haptics.h      -> app/systems/haptics.h
- app/components/input_events.h -> app/systems/input_events.h

Include sites updated across app/systems/, app/ui/, tests/, and
benchmarks/. Design-doc and audio-pipeline-spec references updated to
the new paths.

Validation: 901 Catch2 tests / 2957 assertions pass with zero warnings
under both unity and non-unity Clang -Wall -Wextra -Werror builds.
