+++ app/systems/game_state_terminal_phase_system.cpp
@@ -1,7 +1,7 @@
 #include "all_systems.h"
 #include "game_phase_transition.h"
-#include "../components/audio_events.h"
-#include "../components/haptics.h"
+#include "audio_events.h"
+#include "haptics.h"
 #include "../components/high_score.h"
 #include "../components/rhythm.h"
 #include "../components/scoring.h"
