+++ app/systems/input_dispatcher.cpp
@@ -1,5 +1,5 @@
 #include "input_routing.h"
-#include "../components/input_events.h"
+#include "input_events.h"
 #include "../ui/level_select_controller.h"
 #include <entt/entt.hpp>
 #include <stdexcept>
