+++ app/systems/player_input_system.cpp
@@ -2,9 +2,9 @@
 #include "../components/game_state.h"
 #include "../components/player.h"
 #include "../components/rendering.h"
-#include "../components/input_events.h"
-#include "../components/audio_events.h"
-#include "../components/haptics.h"
+#include "input_events.h"
+#include "audio_events.h"
+#include "haptics.h"
 #include "../components/rhythm.h"
 #include "../constants.h"
 #include "../util/lane_utils.h"
