+++ tests/test_haptics_backend.cpp
@@ -1,6 +1,6 @@
 #include <catch2/catch_test_macros.hpp>
 
-#include "components/haptics.h"
+#include "systems/haptics.h"
 #include "systems/haptics_backend.h"
 
 #include <entt/entt.hpp>
