+++ tests/test_helpers.h
@@ -5,14 +5,14 @@
 #include "components/player.h"
 #include "components/obstacle.h"
 #include "components/input.h"
-#include "components/input_events.h"
+#include "systems/input_events.h"
 #include "components/game_state.h"
 #include "components/scoring.h"
 #include "components/rendering.h"
 #include "components/particle.h"
 #include "components/audio.h"
-#include "components/haptics.h"
-#include "components/audio_events.h"
+#include "systems/haptics.h"
+#include "systems/audio_events.h"
 #include "systems/audio_routing.h"
 #include "entities/settings.h"
 #include "components/rhythm.h"
